handgesturechange/requirements.txt
opencv-python
mediapipe
PyOpenGL
pygame
numpy
PyOpenGL_accelerate
